Improve 'git describe' parsing
The output from 'git describe' varies depending on whether the found tag
is exactly at, or just reachable from, the given revision.

Handle this by running 'git describe' twice for each glob match: first
with '--exact-match' (in which case the output *is* the tag name), and
second without (in which case the output must be parsed to find the tag
name).

Signed-off-by: Tim Waugh <twaugh@redhat.com>